Add doc

+++ test/Vault/in_and_out_functions/Redeem.test.sol
@@ -34,7 +34,7 @@ contract VaultRedeemTest is SimpleVaultTestSetup {
     }
 
     function testPreviewRedeemFee() public {
-        uint256 exitFeeBasisPoints = 100; // 1%
+        uint16 exitFeeBasisPoints = 100; // 1%
         setExitFeeBasisPoint(exitFeeBasisPoints);
 
         // deposit 1000
